* Improved error handling in VariableDeserializer.  * Moved EnvironmentCustomizer to proper package
