+++ src/main/java/io/compprov/core/EnvironmentCustomizer.java
@@ -1,6 +1,4 @@
-package io.compprov;
-
-import io.compprov.core.ComputationEnvironment;
+package io.compprov.core;
 
 public interface EnvironmentCustomizer {
     void customize(ComputationEnvironment environment);
